+++ src/main/java/br/com/edu/ufcg/cccfarma/api/lote/LoteRepositorio.java
@@ -6,15 +6,15 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.edu.ufcg.cccfarma.api.produto.Produto;
 
-public interface LoteRepositorio extends JpaRepository<Lote, Long>{
+public interface LoteRepositorio extends JpaRepository<Lote, Integer>{
 
 	List<Lote> findByProduto(Produto produto);
 
 	List<Lote> findByProdutoCodBarra(String codBarra);
 
-	Lote findByNumeroLote(Long long1);
+	Lote findByNumeroLote(Integer long1);
 
-	Lote getLoteByNumeroLoteAndProdutoCodBarra(Long loteId, String produtoId);
+	Lote getLoteByNumeroLoteAndProdutoCodBarra(Integer loteId, String produtoId);
 	
 	
 }
